mejora sustancial en el manejo de idiomas. Si el navegador del usuario tiene uno de los idiomas soportados lo aplica. Si queremos compartir una URL forzando un idioma se puede usando /es o /en Los textos originales en ingles se toman del HTML
